+++ example/src/app/api/tasks/[id]/(default)/config.ts
@@ -4,7 +4,7 @@ import { ikkanConfig } from "@ikkan/core";
 export const getConfig = ikkanConfig({
   endpoint: ({ id }: { id: string }) => `/api/tasks/${id}`,
   method: "GET",
-  fn: async (_req, { id }) => {
+  fn: async ({segments: { id }}) => {
     return await prisma.task.findUnique({
       where: {
         id: parseInt(id),
@@ -16,7 +16,7 @@ export const getConfig = ikkanConfig({
 export const deleteConfig = ikkanConfig({
   endpoint: ({ id }: { id: string }) => `/api/tasks/${id}`,
   method: "DELETE",
-  fn: async (_req, { id }) => {
+  fn: async ({segments: { id }}) => {
     return await prisma.task.delete({
       where: {
         id: parseInt(id),