+++ example/src/app/api/tasks/[id]/name/config.ts
@@ -8,7 +8,7 @@ export const config = ikkanConfig({
   schema: z.object({
     name: z.string(),
   }),
-  fn: async (_req, { name }, { id }) => {
+  fn: async ({params: { name }, segments: { id }}) => {
     // update the task name
     const task = await prisma.task.update({
       where: {